Add a button to show completed todos
