+++ src/components/TodoList.vue
@@ -5,35 +5,21 @@
       v-model="newTodoTitle"
       placeholder="Enter new todo..."
       @keypress.enter="addTodo">
+    <span class="toggle-complete-button"
+          :class="{'toggle-complete_show': displayComplete}"
+          @click="toggleDisplayComplete">☑</span>
     <div v-if="loadingTodos" class="loading-message">Loading todos...</div>
     <div v-if="failedToLoad" class="failed-loading-message">Failed to load todos.</div>
-    <div
-      v-show="!loadingTodos && !failedToLoad"
-      class="todo-item"
-      v-for="todo in todos"
-      :key="todo.id">
-      <div class="todo-checkbox"
-        :class="{ 'todo-checkbox_checked': todo.complete,
-                  'todo-checkbox_unchecked': !todo.complete }"
-        :value="todo.complete"
-        @click="toggleComplete(todo)">
-          <div v-show="todo.complete">✔</div>
-        </div>
-        <div class="todo-title"
-          :class="{'todo-title_complete': todo.complete}">{{todo.title}}</div>
-      <div class="created-time"
-        :title="getLongDate(todo.created)">
-        {{getShortDate(todo.created)}}
-      </div>
-      <div class="button button-delete" @click="deleteTodo(todo.id)">🗙</div>
+    <div v-show="!loadingTodos && !failedToLoad">
+      <TodoItem v-for="todo in incompleteTodos" :key="todo.id" :todo="todo"/>
+      <TodoItem v-for="todo in completeTodos" v-show="displayComplete" :key="todo.id" :todo="todo"/>
     </div>
   </div>
 </template>
 
 <script>
 import Vue from "vue";
-import { Todo } from "../types/todo";
-import { DateTime } from "luxon";
+import TodoItem from "./TodoItem";
 
 export default Vue.extend({
   data() {
@@ -43,9 +29,21 @@ export default Vue.extend({
       failedToLoad: false,
     };
   },
+  components: {
+    TodoItem,
+  },
   computed: {
     todos: function() {
       return this.$store.state.todos;
+    },
+    incompleteTodos: function() {
+      return this.$store.getters.incompleteTodos;
+    },
+    completeTodos: function() {
+      return this.$store.getters.completeTodos;
+    },
+    displayComplete: function() {
+      return this.$store.state.displayComplete;
     },
   },
   methods: {
@@ -65,18 +63,8 @@ export default Vue.extend({
         });
       }
     },
-    deleteTodo(id) {
-      this.$store.dispatch("deleteTodo", id);
-    },
-    toggleComplete(todo) {
-      const updatedTodo = Object.assign(todo, {complete: !todo.complete});
-      this.$store.dispatch("updateTodo", updatedTodo);
-    },
-    getShortDate(ts_millis) {
-      return DateTime.fromMillis(ts_millis).toFormat("HH:mm d MMM");
-    },
-    getLongDate(ts_millis) {
-      return DateTime.fromMillis(ts_millis).toLocaleString(DateTime.DATETIME_FULL);
+    toggleDisplayComplete() {
+      this.$store.dispatch("toggleDisplayComplete");
     },
   },
   mounted() {
@@ -116,73 +104,21 @@ export default Vue.extend({
   box-shadow: 0px 2px 4px 1px #d4d4d4;
 }
 
-.todo-item {
-  display: flex;
-  max-width: 600px;
-  align-items: center;
-  margin: 5px;
-}
-.todo-item:hover {
-  background-color: #f8f8f8;
-}
-
-.todo-checkbox {
-  flex-shrink: 0;
-  width: 16px;
-  height: 16px;
-  border-radius: 50%;
+.toggle-complete-button {
+  max-width: 80px;
+  font-size: 1.5rem;
+  color: #aaa;
+  background-color: #f0f0f0;
+  border-radius: 10px;
+  padding: 5px 8px;
   cursor: pointer;
-  font-size: 20px;
 }
-.todo-checkbox_unchecked {
-  border: 2px solid #b6b6b6;
-}
-.todo-checkbox_checked {
-  color: #007529;
-  border: 2px solid transparent;
-}
-.todo-checkbox:hover {
-  border: 2px solid transparent;
-}
-.todo-checkbox_unchecked:hover:after {
-  content: "✔";
-  color: #007529;
-}
-
-.todo-title {
-  flex-grow: 1;
-  min-width: 200px;
-  padding: 5px 10px;
-  word-wrap: break-word;
-  text-align: left;
-}
-.todo-title_complete {
-  font-style: italic;
-  text-decoration: line-through;
-  color: #666;
-}
-
-.created-time {
-  font-size: 0.9em;
-  color: #888;
-  min-width: 160px;
-  flex-grow: 0;
-}
-
-.button {
-  cursor: pointer;
-  width: 25px;
-  height: 25px;
-}
-
-.button-delete {
-  flex-shrink: 0;
-  color: #DD0426;
-  background-color: #e2e2e2;
-  border-radius: 3px;
-}
-.button-delete:hover {
+.toggle-complete-button:hover {
   box-shadow: 0px 1px 3px 1px #cbcbcb;
+}
+.toggle-complete_show {
+  color: #007529;
+  background-color: #e5e5e5;
 }
 
 .failed-loading-message {
