+++ src/vuex/store.ts
@@ -1,18 +1,19 @@
 import Vue from "vue";
 import Vuex from "vuex";
-import uuidv4 from "uuid";
 import axios from "axios";
 
 import { Todo } from "../types/todo";
 
 interface State {
   todos: Todo[];
+  displayComplete: boolean;
 }
 interface Mutations {
   LOAD_TODOS: string;
   ADD_TODO: string;
   DELETE_TODO: string;
   UPDATE_TODO: string;
+  TOGGLE_DISPLAY_COMPLETE: string;
 }
 
 const mutations: Mutations = {
@@ -20,19 +21,31 @@ const mutations: Mutations = {
   ADD_TODO: "addTodo",
   DELETE_TODO: "deleteTodo",
   UPDATE_TODO: "updateTodo",
+  TOGGLE_DISPLAY_COMPLETE: "toggleDisplayComplete",
 };
 
 Vue.use(Vuex);
 
 const initialState: State = {
   todos: [],
+  displayComplete: false,
 };
 
 const store = new Vuex.Store({
   state: initialState,
 
+  getters: {
+    incompleteTodos: (state) => {
+      return state.todos.filter((todo) => !todo.complete);
+    },
+    completeTodos: (state) => {
+      return state.todos.filter((todo) => todo.complete);
+    },
+  },
+
   mutations: {
     addTodo(state, data) {
+      // Filter out duplicates that have already been loaded
       state.todos = state.todos.filter((todo) => data.id !== todo.id);
       state.todos.push(data);
     },
@@ -46,6 +59,9 @@ const store = new Vuex.Store({
         }
         return todo;
       });
+    },
+    toggleDisplayComplete(state) {
+      state.displayComplete = !state.displayComplete;
     },
   },
 
@@ -104,6 +120,9 @@ const store = new Vuex.Store({
         .catch((err) => {
           console.error(err);
         });
+    },
+    toggleDisplayComplete({ commit }) {
+      commit(mutations.TOGGLE_DISPLAY_COMPLETE);
     },
   },
 });
